removed uint casting

+++ philo/utils.c
@@ -6,7 +6,7 @@
 /*   By: ketrevis <ketrevis@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/23 18:23:45 by ketrevis          #+#    #+#             */
-/*   Updated: 2024/01/24 14:52:22 by ketrevis         ###   ########.fr       */
+/*   Updated: 2024/01/26 11:56:15 by ketrevis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	get_curr_time(t_philo *philo)
 	pthread_mutex_lock(&philo->args->time_mutex);
 	gettimeofday(&tv, NULL);
 	pthread_mutex_unlock(&philo->args->time_mutex);
-	return ((((uint)tv.tv_sec) * 1000) + (tv.tv_usec / 1000));
+	return (((tv.tv_sec) * 1000) + (tv.tv_usec / 1000));
 }
 
 void	init_odd_philo(t_philo *philo)
